Format logic and parenthesized patterns (#1380)
* Change CodeWriter to make the indentation stack explicit.

Prior to this change, CodeWriter maintained an implicit stack where
each Piece being formatted has its own indentation level which it can
mutate. When the Piece is done, whatever amount of relative indentation
it set is discarded.

This commit replaces that with an explicit stack that Pieces manipulate
by pushing and popping indentation levels onto it. Pieces that don't
care about indentation don't affect the stack at all. Pieces that need
multiple levels of indentation (like constructor initializers) can push
and pop multiple times.

* Use an explicit stack for "allow newlines".

As with the previous commit, instead of each Piece implicitly having
it's own context where this is stored, require pieces to explicitly
push and pop regions where newlines are allowed or disallowed.

With that change, there's almost nothing left in the _Options class, so
eliminate that too.

* Refactor ChainPiece.format().

I think it's clearer to handle each state separately.

* Format logic patterns.

In order to get infix patterns (and split chain constant patterns, and
others) indenting correctly in if-case statements, I had to introduce
a notion of "collapsible" indent.

In this PR, it's only used for the indentation after if-case, but it
might be useful elsewhere (for example, with `=>` function bodies).

* Fix incorrect indent stack popping.