Format metadata annotations.
This handles metadata everywhere except on for loop variables. I'll do
that later after patterns in for loops are supported.

Since there were some metadata tests on pattern variable declarations,
this also includes a trivial implementation of that. There's a TODO to
revisit it and add real tests, but it's enough for the metadata tests.

Figuring out where to put the metadata-handling code was a little tricky
since so many places in the language can have metadata. I ultimately
decided to put it into AdjacentBuilder. Every place that was working
with metadata also had one of those builders in play, so that made it
easy to add a simple call to `b.metadata()` and get the metadata
working.

It does mean that AdjacentBuilder isn't strictly about adjacent pieces
anymore, but I think that's OK. In practice, it's used mostly as a
general "build pieces out of smaller things" builder.