Format pattern variable declarations, including in for loops. (#1392)
Format pattern variable declarations, including in for loops.

This required pretty significantly revamping AssignPiece since now
there can be complex syntax on the left side of the `=` too. It went
through several iterations but I think I managed to simplify it down
to something reasonable that still passes all of the tests.

This PR doesn't do pattern assignment, but that should hopefully be a
trivial change after this.