Add more tests for type arguments in object patterns and remove a TODO. (#1401)
* Add more tests for type arguments in object patterns and remove a TODO.

I'm honestly not sure what the TODO was even trying to say. But I think
the formatter behaves like we want here so there's nothing TODO as far
as I can tell. But this corner did seem a little under tested, so added
some more.

* Fix indentation collapsing.

A block indent shouldn't be able to partially eat an expression-sized
collapsible indentation. The only time we want to actually collapse is
when we have two equivalent expression-sized indentations.