Force switch expressions to split if they contain a line comment.
The fix here is pretty hacky, but this is in the old formatter which
will be going away hopefully relatively soon. I honestly don't know how
to fix it more elegantly and I'm a little worried that any more systemic
fix will cause knock-on bugs since I don't have the old code loaded in
my head as well as I used to.

Fixes #1404.