Change the Piece structure for Object patterns.
This change is mainly to ensure that every outermost Piece for a pattern
that can be block-formatted is a ListPiece. That will be needed in a
later commit so that when a pattern appears to the left of an `=`, we
can constrain it to block split sometimes.

But this change also slightly changes the formatting of object patterns
in a way that I think looks better: if the type arguments in an object
pattern split, it forces the fields to split too.