Move the operator in AssignPiece to a separate child piece.
Prior to this change, the operator was directly attached to either the
LHS piece (for ` =`) or the RHS (for `in `). This pulls it out to its
own piece. The formatting is unchanged by this commit.

In a future commit, we'll use this to be able to constrain the LHS
child piece without the operator getting in the way.