Tweak the heuristics on which expressions allow block formatting.
I was trying out the new formatting on Flutter and noticed a few issues:

1. Immediately-invoked functions in asserts are common.

Prior to this change, these were not block format candidates, giving
you:

```dart
assert(
  () {
    // Some slow computation...
  }(),
  'Message.',
);
```

With this change, they're treated like (uninvoked) function expressions:

```dart
assert(() {
  // Some slow computation...
}(), 'Message.');
```

2. Formatting large deep widget trees looks weird.

Prior to this change, a function call inside an argument list can be
treated like a block argument even if there are other arguments. That
tends to pack things in strangely in widget code, like:

```dart
SizedBox(height: 1000.0, width: double.infinity, child: Column(
  children: <Widget>[Scrollbar(key: key1, child: SizedBox(
    height: 300.0,
    width: double.infinity,
    child: SingleChildScrollView(key: innerKey, child: const SizedBox(key: Key(
      'Inner scrollable',
    ), height: 1000.0, width: double.infinity)),
  ))],
));
```

With this change, a function or method call inside an argument list is
only block formatted if there are no other arguments. In other words,
you can block format a function call if the outer call is a pure wrapper
around the call, but not otherwise. That leads to:

```dart
SizedBox(
  height: 1000.0,
  width: double.infinity,
  child: Column(children: <Widget>[Scrollbar(
    key: key1,
    child: SizedBox(
      height: 300.0,
      width: double.infinity,
      child: SingleChildScrollView(
        key: innerKey,
        child: const SizedBox(
          key: Key('Inner scrollable'),
          height: 1000.0,
          width: double.infinity,
        ),
      ),
    ),
  )]),
);
```

I think that looks a lot nicer.

3. Deep, complex argument lists are combinatorially slow.

We have an optimization to eagerly force an argument list to split if
its contents clearly won't fit. But that optimization ignores the size
of a potential block argument since the contents of that can be wrapped
across multiple lines without forcing the outer argument list to split.

A consequence of that is that big deeply nested function calls end up
slow if it happens that each one only has a single block-formattable
call. That turns out to be very common in Flutter code where you have a
deep chain of nested widgets along with a bunch of other arguments.

The above change to disallow block formatting function calls if there
are any other arguments conveniently fixes that performance issue in
most cases. (Though I expect we will want to do more work here to handle
pathological cases.)