Fix a crash and some formatting bugs in cascades.
I did a test run of the new formatter on the Flutter repo and it hit
some crashes in cascades where a cascade section contains a method
chain like `foo..a.b`.

While investigating that, I realized that the existing code for
formatting cascades tries to reuse the regular method chain formatting
code entirely unnecessarily.

This fixes the crash, and the formatting, and is simpler for handling
cascades.