Force list, set, and maps to split if they are nested.
If a list, set, or map literal contains (directly or indirectly) another
list, set, or map literal, then force the outer one to split, even if
it would otherwise fit.

We don't do this for record expressions, or any kind of pattern because
those tend to be smaller and look better when kept closely packed.

The behavior here is exactly the same as in the current formatter.