Test that benchmarks produce their expected output.
The benchmark expectations weren't being run as part of the tests which
meant that it was easy to break the benchmarks without realizing it. A
recent fix to not block format collection literal elements broke a
couple of the benchmarks.

This fixes them. It also ensures they won't break again by having the
format tests also test all of the benchmarks.

And it adds support to the benchmark runner to run all the benchmarks
in one go.