Support leading indentation in the new style.
The formatter library API allows you to pass in a number of leading
indent spaces that will be applied to all lines of the output. This is
used internally for the regression tests (which often need to preserve
the indentation of the surrounding context of the original failing
code). It's also supported by the API so that some editors which support
"format just this selection" can pass in a leading indent of the
surrounding context.

This passes that through to the new formatter.

(It wasn't caught by tests because the API and command-line tests are
currently not forked to run the new style too.)