Preserve newlines (or lack of) in collections with line comments. (#1420)
Preserve newlines (or lack of) in collections with line comments.

Large collection literals often have some internal structure that the
author wants to highlight by having multiple elements on one line, sort
of like a matrix.

The old style has a sort of hacky but in practice useful heuristic to
enable that: if a collection literal contains a line comment, then the
newlines (or lack of) between pairs of elements are preserved:

https://github.com/dart-lang/dart_style/wiki/FAQ#why-does-the-formatter-mess-up-my-collection-literals

This implements that same heuristic in the new style. The behavior is
slightly different in the new style. In the old style, an argument list
can be split in a variety of ways:

```dart
// 1. All on one line:
f(argument1, argument2, argument3, argument4)

// 2. Split into two lines at after any particular argument:
f(argument1, argument2, argument3,
    argument4)
f(argument1, argument2,
    argument3, argument4)
f(argument1,
    argument2, argument3, argument4)

// 3. Split every argument into its own line:
f(
    argument1,
    argument2,
    argument3,
    argument4)
```

It uses that same logic for the elements on a single line in a
collection formatted with this rule. That kicks in if the authored code
has all of the elements on one line but they don't actually fit. In that
case, the formatter will still split them to try to fit. With the old
style, it may use any of the above approaches to try to fit.

In the new style, argument lists don't support splitting between an
arbitrary pair of elements like (2) above. So you just get:

```dart
// 1. All on one line:
f(argument1, argument2, argument3, argument4)

// 3. Split every argument into its own line:
f(
  argument1,
  argument2,
  argument3,
  argument4,
)
```

Likewise for collection literals containing a line comment. If all of
the elements on a single line don't fit, it will split all of them onto
their own lines instead of trying to split just between one pair of
them.

Longer-term, I intend to support opting regions of code out of
formatting completely. I expect users will use that for large
matrix-like collections where they want to control both line splitting
and aligning things into columns. So the specific behavior of this hack
won't matter as much.

But it *is* a widely relied on rule today, so this part preserves it.

Co-authored-by: Nate Bosch <nbosch@google.com>