Reorganize the sources and tests to split out the short and tall files.
We're going to be supporting both the short and tall styles for at
least a couple of Dart releases, so having all of the directories and
files for both jumbled together is probably silly.

This moves all of the short style test cases under test/short/ and the
tall test cases to test/tall/.

It takes all of the source files and directories under src/ that are
specific the short style and puts them under src/short/. The source
files that are used by both styles or only by the tall style stay where
they are.

There are no other changes (except a couple of tweaks to the test util
code since now we can tell whether a test is short or tall style based
on the file path).