Tweak the block formatting rules to disallow `=>` function expressions. (#1422)
Tweak the block formatting rules to disallow `=>` function expressions.

When I first added support for block formatting arguments in the new
style, I allowed function expressions with `=>` bodies if they had
parameters with the idea that the parameter list itself would get the
block-like formatting, as in:

```dart
function((
  parameter1,
  parameter2,
) => body);
```

I'm honestly not sure why I thought that was a good idea. The old
formatter doesn't allow that. I don't think it looks particularly good.
And, worst of all, the formatter doesn't actually *require* the
parameter list to be the place where the split occurs, so the current
implementation allows unwanted output like:

```dart
function((parameter) => operand +
    anotherOperand));
```

This is definitely not intended.

I've started migrating regressions tests and in the examples I've seen
where this rule comes into play, I think it looks worse to allow block
formatting here. So this PR changes it to not allow block formatting
any `=>` body function expression.