Normalize newlines at the end of the tests.
Most test files don't end with a newline. That's admittedly unusual for
text files, but I did it that way to avoid confusion about whether the
last test in the file is itself required to output a trailing newline.

In practice, it seems to work fine whether or not there is a trailing
newline now? Anyway, for consistency's sake, remove all the trailing
newlines. That way, the test updater doesn't produce spurious diffs on
these files when it tries to remove them.