Clean up a TODO.
The strangely-located "### TODO: Clean up." in the middle of this file
was supposed to be a reminder to me to clean up the test descriptions
in this file before I sent it out for review but I guess I forgot. :)

Doing it now to make the test cleaner and because the "###" comment is
in a place where comments are no longer allowed.