Update tool/test_updater.dart to handle the tall style and test changes. (#1426)
Update tool/test_updater.dart to handle the tall style and test changes.

This internal tool is used to automatically update the expectations for
test files based on the current behavior of the formatter. This PR gets
it working with all of the changes that have happened since the new
style. Those are:

- Use the tall style when updating a tall/ test.

- Enable the inline-class and macros experiments generating the output.
  The format test runner itself always enables those.

- Preserve `###` comments. This is tricky if we allow those comments to
  appear anywhere in a test since it might not be clear where they
  should be reinserted in an updated test expectation. To avoid that,
  this change now only allows `###` comments to appear at the top of the
  test file, or at the beginning of an actual or expected section. All
  of the existing comments already fit that restriction, except for one
  which was never supposed to be committed. I'll fix that in a later
  commit.

- Handle skipping new tall style tests that contain selections or
  Unicode escapes. The test updater doesn't know how to update those
  tests.

Also, I generally made the output a little nicer.

* Normalize newlines at the end of the tests.

Most test files don't end with a newline. That's admittedly unusual for
text files, but I did it that way to avoid confusion about whether the
last test in the file is itself required to output a trailing newline.

In practice, it seems to work fine whether or not there is a trailing
newline now? Anyway, for consistency's sake, remove all the trailing
newlines. That way, the test updater doesn't produce spurious diffs on
these files when it tries to remove them.

* Fix a typo in a test expectation.

This should have always been ">>>", but the TestFile parser happens to
always treat the first line of the test file (after the "|" or "###"
lines) as the beginning of a test, even if it doesn't start with ">>>".

Running this file through the test updater normalized it to the correct
syntax.

Also, clean up a TODO.

The strangely-located "### TODO: Clean up." in the middle of this file
was supposed to be a reminder to me to clean up the test descriptions
in this file before I sent it out for review but I guess I forgot. :)

Doing it now to make the test cleaner and because the "###" comment is
in a place where comments are no longer allowed.