Indent comments inside switch statement cases.
Before this fix, line comments inside a switch case body would be
indented to align with the case, not the body. This aligns them with
the case body, except for comments inside a case but after the body.
Those are still aligned to the case level because most of the time when
comments appear at the end of a case body, they describe the subsequent
case, like:

```dart
switch (obj) {
  case 1:
    body;

  // Comment here describes next case:
  case 2:
    body;
}
```