Make comments into separate pieces.
This is the big commit. It's hard to break into smaller pieces since
everything is so intertwined.

This, I believe, makes PieceWriter less mutable and easier to follow.
Instead of a `_currentPiece` TextPiece which is continually being
mutated by having comment text and token lexemes stuffed into it, each
TextPiece is almost immutable: We basically create one in one operation
and then it's done except that hanging comments may be added later.

This change slightly modifies the formatting of comments when they
appear in a couple of unusual places. The previous formatting was pretty
arbitrary. There are even comments in the tests acknowledging that it
looks weird. So I'm OK with that.

I also added some more tests around comments in enums because that's a
unique corner case where we discard a trailing comma and have to be
careful not to mix up comments.