Make subclasses of TextPiece for code and comments.
In order for comments to be their own pieces, it means TextPiece needs
to have child pieces so that a piece for a single token can have
comments attached to it.

The previous commit does that, but it's not optimal: A TextPiece for a
comment will never have any child pieces because we don't attach
comments to comments. Conversely, a TextPiece for a token will never use
_trailingWhitespace since that's only for ensuring whitespace around
comments.

This solves that by making separate subclasses of TextPiece for each
use case.