Don't indent split binary operands in an assignment-like context. (#1431)
Don't indent split binary operands in an assignment-like context.

Binary operators indent their subsequent operands in most places:

```dart
function(
  operand1 +
      operand2 +
      operand3,
);
```

But it's gratuitous on contexts where a split between the operands will also lead to a split before the operand. In that case, it looks nicer to align the operands:

```dart
variable =
    operand1 +
    operand2 +
    operand3;
```

The places where this can occur are "assignment-like": the right-hand side of `=`, `:`, or `=>`.

This is another example where the relationship between a parent and child AST node affects how they are formatted. We don't have a great systematic way of modeling those relationships in the codebase right now. Sometimes, we have the code building the parent Piece peek at the child's AST node type and do something different with the child based on that. Sometimes, the child calls `.parent` on itself and peeks at the parent type. We don't generally look at the type of the Piece of the child because we don't consider that to be reliably encoding what kind of original syntax it came from (though we could).

This takes yet another approach where we allow a parent AST node to pass in a "context" when visiting one of its children. A child can then look at that context to tweak its formatting. I'm not sure if I love the approach, but I thought I'd give it a try and see what you think.