Rework how comments are represented in the Piece model (#1432)
I've been frustrated with how the new formatter represents comments for a while. Jamming them inside the _lines list in TextPiece sort of works, but it means we lose track of what's a comment and what isn't. It means PieceWriter and TextPiece are more mutable than I'd like, which is I think a large part of why the logic of those classes is hard to follow.

I was trying to fix some bugs that I noticed when migrating the regression tests. For example:

value =
    // comment
    a + b;
Currently, the formatter outputs:

value =
    // comment
    a +
        b;
That's because the // comment is part of the TextPiece for a, so it thinks there's a newline inside the first operand to + and thus the + splits.

Trying to fix that with the current comment handling was really hard. So I finally dug in and tried to come up with a cleaner way to model comments. This PR does that. The basic idea is that we make each comment a separate piece. That in turn implies that TextPieces for non-comment code needs to support child pieces for the comments attached to them.

It doesn't, like, radically improve comment handling. (It also doesn't fix the above bug yet.) But I do think the resulting code is cleaner, and I have a separate fix for the above bug that builds on it that I think would have been harder to do without.

Also, I have a goal to do an optimization where we coalesce adjacent TextPiece and SpacePieces when we can to produce a smaller, chunkier piece tree. In my informal tests, it measurably improves perf. But I've never been able to get a fully working implementation because there are so many corner cases around the mutability in PieceWriter that it was a bug farm. I hope this PR will make that more feasible.