Add some regression tests from the old prototype. (#1435)
When I was working on the prototype implementation, I ran it on a bunch
of codebases. I got it working as well as I could but I sometimes noted
cases where the prototype did the wrong thing. I held onto them to make
sure that the real implementation is better.

These are most of those cases. I still have a few where the new
formatter needs a tweak or two to get it right, but these ones are
already doing the right thing.